refactor: add the following:
- remove duplicated lower-case components
- remove unused `EnableAccount` and `Drawer` component
- fix components naming
