--- components/button/EnableAccount.tsx
@@ -1,5 +1,0 @@
-import { Button } from "@chakra-ui/react";
-
-export const EnableAccount = () => {
-  return <Button>Enable Account</Button>;
-};
