+++ pages/address/[address].tsx
@@ -13,14 +13,14 @@ import {
 } from "@chakra-ui/react";
 import { ArrowForwardIcon } from "@chakra-ui/icons";
 import Link from "next/link";
-import LinkLayout from "../../components/AppLayout/linkLayout";
+
 import { connectToDatabase } from "../../util/mongodb";
 
 const Address = (props: any) => {
   const { txs, address } = props;
 
   return (
-    <LinkLayout>
+    <>
       <Breadcrumb
         ml="20px"
         mb="5px"
@@ -98,7 +98,7 @@ const Address = (props: any) => {
           })}
         </Tbody>
       </Table>
-    </LinkLayout>
+    </>
   );
 };
